+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.6.0] - 2026-03-24
+
 ### TOFcam635
 - added setIlluminatorSegments command